Removing some more msvc warnings.

+++ plugins/mtoa/render/RenderSwatch.cpp
@@ -361,7 +361,7 @@ MStatus CRenderSwatchGenerator::AssignNode(AtNode* arnoldNode, CNodeTranslator* 
    
    AiNodeDeclare(options, "frame", "constant FLOAT");
    MTime ct = MAnimControl::currentTime();
-   AiNodeSetFlt(options, "frame", ct.value());
+   AiNodeSetFlt(options, "frame", (float)ct.value());
 
    // If we are swatching an environment (background) shader
    if (m_swatchClass == SWATCH_ENVIRONMENT)
